More casts

+++ source/app/attributes/enrichmentcalculator.cpp
@@ -114,7 +114,7 @@ EnrichmentTableModel::Table EnrichmentCalculator::overRepAgainstEachAttribute(co
         for(auto& attributeValueB : u::keysFor(attributeValueEntryCountBTotal))
         {
             EnrichmentTableModel::Row row(EnrichmentTableModel::Results::NumResultColumns);
-            command.setProgress(progress * 100U / iterations);
+            command.setProgress(static_cast<int>(progress * 100U / iterations));
             progress++;
 
             n = graphModel->graph().numNodes();
